// src/Models/ECustomKey.ts
export const enum ECustomKey {
	Thumbstick1Up = "Thumbstick1Up",
	Thumbstick2Up = "Thumbstick2Up",

	Thumbstick1Down = "Thumbstick1Down",
	Thumbstick2Down = "Thumbstick2Down",

	Thumbstick1Left = "Thumbstick1Left",
	Thumbstick2Left = "Thumbstick2Left",

	Thumbstick1Right = "Thumbstick1Right",
	Thumbstick2Right = "Thumbstick2Right",

	MouseWheelUp = "MouseWheelUp",
	MouseWheelDown = "MouseWheelDown",

	MouseLeft = "MouseLeft",
	MouseRight = "MouseRight",
	MouseDown = "MouseDown",
	MouseUp = "MouseUp",
}